feat(autoformat): Simplify autoformatting with Neoformat
This commit simplifies the autoformatting setup in Neovim by replacing
the previous custom implementation with Neoformat. The change includes
the creation of a new Neoformat configuration file and the modification
of the autocmds.lua file to use Neoformat for autoformatting on save.
This should make the autoformatting process more efficient and easier
to maintain.